Cleaned things up.
