+++ src/Module.Server/MapRotation/ds_config_crpg_dtv.txt
@@ -1,4 +1,4 @@
-﻿ServerName cRPG - Test DTV3 - c-rpg.eu
+﻿ServerName cRPG - Test DTV2 - c-rpg.eu
 GameType cRPGDTV
 GamePassword yoyo
 AllowPollsToKickPlayers True
