+++ src/Module.Server/Modes/Dtv/CrpgDtvGameMode.cs
@@ -149,7 +149,7 @@ internal class CrpgDtvGameMode : MissionBasedMultiplayerGameMode
                 new DrowningBehavior(),
                 new PopulationBasedEntityVisibilityBehavior(lobbyComponent),
                 new CrpgCommanderBehaviorServer(),
-                new EquipmentChestTimeoutBehavior(rewardServer, crpgClient),
+                new EquipmentChestTimeoutBehavior(spawnBehaviour),
 
 #else
                 new MultiplayerAchievementComponent(),
